rename: strategy -> provider
